changed file names + folder name + modified checks in v2
