+++ FunctionLevelLipsDetection/SkinExtraction.py
@@ -1,7 +1,5 @@
 import cv2
 import numpy as np
-
-
 # ----------------------------------------------------------------------------
 # function used to exract skin from image
 # input: image
@@ -45,10 +43,7 @@ def segmentSkin(img):
     holes_filled_skin_image = fill_holes(skin_image)
     holes_filled_skin_image = holes_filled_skin_image.astype(np.uint8)
     holes_filled_skin_image = cv2.cvtColor(holes_filled_skin_image, cv2.COLOR_RGB2GRAY)
-    # cv2.imshow('skin_image', holes_filled_skin_image)
     return holes_filled_skin_image
-
-
 # ----------------------------------------------------------------------------
 # function used to support segmentSkin function by filling holes in the skin mask from image
 # input: skin mask
@@ -57,8 +52,6 @@ def fill_holes(img):
     kernel = cv2.getStructuringElement(cv2.MORPH_ELLIPSE, (5, 5))
     res = cv2.morphologyEx(img, cv2.MORPH_OPEN, kernel)
     return res
-
-
 # ----------------------------------------------------------------------------
 # function used to exract skin from image
 # input: image, skin mask
@@ -77,7 +70,7 @@ def extractSkin(img, skinMask):
     contour_info = []
     contours, _ = cv2.findContours(edges, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)
     if len(contours) == 0:
-        return []
+        return [], False
     cntf = []
     for c in contours:
         contour_info.append((
@@ -86,12 +79,9 @@ def extractSkin(img, skinMask):
             cv2.contourArea(c),
         ))
     cv2.drawContours(img, contours, -1, 255, 3)
-    # cv2.imshow('contours', img) 
-    # cv2.waitKey(0)
     contour_info = sorted(contour_info, key=lambda c: c[2], reverse=True)
     # --------------------------
     max_contour = contour_info[0]
-
     mask = np.zeros(edges.shape)
     cv2.fillConvexPoly(mask, max_contour[0], (255))
 
@@ -106,4 +96,4 @@ def extractSkin(img, skinMask):
 
     masked = (mask_stack * img) + ((1 - mask_stack) * MASK_COLOR)
     masked = (masked * 255).astype('uint8')
-    return masked
+    return masked, True
